Change return types of average grade methods in `Class` from `int` to `

+++ Class.cpp
@@ -229,11 +229,11 @@ int Class::get_id() const {
     return id;
 }
 
-int Class::get_average_grade() const {
+float Class::get_average_grade() const {
     throw std::runtime_error("Method not implemented");
 }
 
-int Class::get_average_grade_from_subject(const std::string &subject) const {
+float Class::get_average_grade_from_subject(const std::string &subject) const {
     throw std::runtime_error("Method not implemented");
 }
